rpi/ML-model/requirements.txt
tensorflow==2.14.0
keras==2.14.0
nats-py==2.3.1
pandas==2.1.1
numpy==1.23.5
